# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mokua Literary Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Author {
  id: number;
  name: string;
  bio: string;
  avatarUrl: string | null;
  twitterHandle: string | null;
  instagramHandle: string | null;
  website: string | null;
  postCount: number;
}

export interface AuthorDetail {
  id: number;
  name: string;
  bio: string;
  avatarUrl: string | null;
  twitterHandle: string | null;
  instagramHandle: string | null;
  website: string | null;
  postCount: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string | null;
  postCount: number;
}

export interface Tag {
  id: number;
  name: string;
  slug: string;
  postCount: number;
}

export type PostSummaryContentType =
  (typeof PostSummaryContentType)[keyof typeof PostSummaryContentType];

export const PostSummaryContentType = {
  story: "story",
  poem: "poem",
  essay: "essay",
  review: "review",
} as const;

export interface PostSummary {
  id: number;
  title: string;
  slug: string;
  excerpt: string | null;
  contentType: PostSummaryContentType;
  featuredImageUrl: string | null;
  readingTime: number;
  publishedAt: string;
  author: Author;
  category: Category;
  tags: Tag[];
}

export type PostDetailContentType =
  (typeof PostDetailContentType)[keyof typeof PostDetailContentType];

export const PostDetailContentType = {
  story: "story",
  poem: "poem",
  essay: "essay",
  review: "review",
} as const;

export interface PostDetail {
  id: number;
  title: string;
  slug: string;
  excerpt: string | null;
  contentType: PostDetailContentType;
  content: string;
  featuredImageUrl: string | null;
  readingTime: number;
  publishedAt: string;
  author: Author;
  category: Category;
  tags: Tag[];
  seoTitle: string | null;
  seoDescription: string | null;
  seoOgImageUrl: string | null;
}

export interface PostListResponse {
  posts: PostSummary[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
  nextCursor: string | null;
}

export type CreateSubmissionRequestContentType =
  (typeof CreateSubmissionRequestContentType)[keyof typeof CreateSubmissionRequestContentType];

export const CreateSubmissionRequestContentType = {
  story: "story",
  poem: "poem",
  essay: "essay",
  review: "review",
} as const;

export interface CreateSubmissionRequest {
  /**
   * @minLength 2
   * @maxLength 100
   */
  submitterName: string;
  submitterEmail: string;
  /**
   * @minLength 2
   * @maxLength 200
   */
  title: string;
  contentType: CreateSubmissionRequestContentType;
  /** @minLength 50 */
  content: string;
  honeypot?: string;
}

export interface SubmissionResponse {
  id: number;
  message: string;
}

export interface NewsletterSubscribeRequest {
  email: string;
  name?: string | null;
}

export interface ContactMessageRequest {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  email: string;
  /**
   * @minLength 2
   * @maxLength 200
   */
  subject: string;
  /** @minLength 10 */
  message: string;
  honeypot?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  details: string | null;
}

export type ListPostsParams = {
  page?: number;
  limit?: number;
  category?: string;
  contentType?: ListPostsContentType;
  tag?: string;
  author?: number;
  cursor?: string;
};

export type ListPostsContentType =
  (typeof ListPostsContentType)[keyof typeof ListPostsContentType];

export const ListPostsContentType = {
  story: "story",
  poem: "poem",
  essay: "essay",
  review: "review",
} as const;

export type SearchPostsParams = {
  q: string;
  category?: string;
  contentType?: SearchPostsContentType;
  page?: number;
  limit?: number;
};

export type SearchPostsContentType =
  (typeof SearchPostsContentType)[keyof typeof SearchPostsContentType];

export const SearchPostsContentType = {
  story: "story",
  poem: "poem",
  essay: "essay",
  review: "review",
} as const;

export type GetRelatedPosts200 = {
  posts: PostSummary[];
};

export type ListCategories200 = {
  categories: Category[];
};

export type ListAuthors200 = {
  authors: Author[];
};

export type GetAuthorPostsParams = {
  page?: number;
  limit?: number;
};

export type ListTags200 = {
  tags: Tag[];
};