# lib/api-zod/src/generated/types/authorDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mokua Literary Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthorDetail {
  id: number;
  name: string;
  bio: string;
  avatarUrl: string | null;
  twitterHandle: string | null;
  instagramHandle: string | null;
  website: string | null;
  postCount: number;
}