# lib/api-zod/src/generated/types/createSubmissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mokua Literary Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSubmissionRequestContentType } from "./createSubmissionRequestContentType";

export interface CreateSubmissionRequest {
  /**
   * @minLength 2
   * @maxLength 100
   */
  submitterName: string;
  submitterEmail: string;
  /**
   * @minLength 2
   * @maxLength 200
   */
  title: string;
  contentType: CreateSubmissionRequestContentType;
  /** @minLength 50 */
  content: string;
  honeypot?: string;
}