# lib/api-zod/src/generated/types/newsletterSubscribeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mokua Literary Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscribeRequest {
  email: string;
  name?: string | null;
}