# lib/api-zod/src/generated/types/postListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mokua Literary Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PostSummary } from "./postSummary";

export interface PostListResponse {
  posts: PostSummary[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
  nextCursor: string | null;
}